requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
face-recognition==1.3.0
numpy==1.26.2
opencv-python==4.8.1.78
sqlalchemy==2.0.23
pillow==10.1.0
python-multipart==0.0.6
pyodbc==4.0.39 
face-recognition-models==0.3.0
